initial attempt of converting nullaway to checker
