+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/Annotator.java
@@ -29,6 +29,7 @@ import edu.ucr.cs.riple.core.cache.TargetModuleCache;
 import edu.ucr.cs.riple.core.cache.downstream.DownstreamImpactCache;
 import edu.ucr.cs.riple.core.cache.downstream.DownstreamImpactCacheImpl;
 import edu.ucr.cs.riple.core.cache.downstream.VoidDownstreamImpactCache;
+import edu.ucr.cs.riple.core.checkers.deserializers.nullaway.NullAway;
 import edu.ucr.cs.riple.core.evaluators.BasicEvaluator;
 import edu.ucr.cs.riple.core.evaluators.CachedEvaluator;
 import edu.ucr.cs.riple.core.evaluators.Evaluator;
@@ -38,15 +39,11 @@ import edu.ucr.cs.riple.core.evaluators.suppliers.TargetModuleSupplier;
 import edu.ucr.cs.riple.core.injectors.AnnotationInjector;
 import edu.ucr.cs.riple.core.injectors.PhysicalInjector;
 import edu.ucr.cs.riple.core.metadata.field.FieldInitializationStore;
-import edu.ucr.cs.riple.core.metadata.index.Error;
 import edu.ucr.cs.riple.core.metadata.index.Fix;
 import edu.ucr.cs.riple.core.util.Utility;
 import edu.ucr.cs.riple.injector.changes.AddAnnotation;
 import edu.ucr.cs.riple.injector.changes.AddMarkerAnnotation;
-import edu.ucr.cs.riple.injector.changes.AddSingleElementAnnotation;
 import edu.ucr.cs.riple.injector.location.OnField;
-import edu.ucr.cs.riple.injector.location.OnParameter;
-import java.util.Objects;
 import java.util.Set;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
@@ -137,8 +134,8 @@ public class Annotator {
       }
     }
 
-    if (config.forceResolveActivated) {
-      forceResolveRemainingErrors();
+    if (config.suppressRemainingErrors) {
+      suppressRemainingErrors();
     }
 
     System.out.println("\nFinished annotating.");
@@ -238,135 +235,9 @@ public class Annotator {
    *       {@code @SuppressWarnings("NullAway")}.
    * </ul>
    */
-  private void forceResolveRemainingErrors() {
-    // Collect regions with remaining errors.
-    Utility.buildTarget(config);
-    Set<Error> remainingErrors =
-        Utility.readErrorsFromOutputDirectory(config, config.targetModuleContext);
-    Set<Fix> remainingFixes =
-        Utility.readFixesFromOutputDirectory(config, config.targetModuleContext);
-    // Collect all regions for NullUnmarked.
-    // For all errors in regions which correspond to a method's body, we can add @NullUnmarked at
-    // the method level.
-    Set<AddAnnotation> nullUnMarkedAnnotations =
-        remainingErrors.stream()
-            // find the corresponding method nodes.
-            .map(
-                error -> {
-                  if (error.getRegion().isOnCallable()
-                      &&
-                      // We suppress initialization errors reported on constructors using
-                      // @SuppressWarnings("NullAway.Init"). We add @NullUnmarked on constructors
-                      // only for errors in the body of the constructor.
-                      !error.isInitializationError()) {
-                    return config
-                        .targetModuleContext
-                        .getMethodRegistry()
-                        .findMethodByName(error.encClass(), error.encMember());
-                  }
-                  // For methods invoked in an initialization region, where the error is that
-                  // `@Nullable` is being passed as an argument, we add a `@NullUnmarked` annotation
-                  // to the called method.
-                  if (error.messageType.equals("PASS_NULLABLE")
-                      && error.isSingleFix()
-                      && error.toResolvingLocation().isOnParameter()) {
-                    OnParameter nullableParameter = error.toResolvingParameter();
-                    return config
-                        .targetModuleContext
-                        .getMethodRegistry()
-                        .findMethodByName(nullableParameter.clazz, nullableParameter.method);
-                  }
-                  return null;
-                })
-            // Filter null values from map above.
-            .filter(Objects::nonNull)
-            .map(node -> new AddMarkerAnnotation(node.location, config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet());
-
-    // For errors within static initialization blocks, add a @NullUnmarked annotation on the
-    // enclosing class
-    nullUnMarkedAnnotations.addAll(
-        remainingErrors.stream()
-            .filter(
-                error ->
-                    error.getRegion().isOnInitializationBlock()
-                        && !error.getRegion().isInAnonymousClass())
-            .map(
-                error ->
-                    new AddMarkerAnnotation(
-                        config.targetModuleContext.getLocationOnClass(error.getRegion().clazz),
-                        config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet()));
-    injector.injectAnnotations(nullUnMarkedAnnotations);
-    // Update log.
-    config.log.updateInjectedAnnotations(nullUnMarkedAnnotations);
-
-    // Collect suppress warnings, errors on field declaration regions.
-    Set<OnField> fieldsWithSuppressWarnings =
-        remainingErrors.stream()
-            .filter(
-                error -> {
-                  if (!error.getRegion().isOnField()) {
-                    return false;
-                  }
-                  if (error.messageType.equals("PASS_NULLABLE")) {
-                    // It is already resolved with @NullUnmarked selected above.
-                    return false;
-                  }
-                  // We can silence them by SuppressWarnings("NullAway.Init")
-                  return !error.isInitializationError();
-                })
-            .map(
-                error ->
-                    config
-                        .targetModuleContext
-                        .getFieldRegistry()
-                        .getLocationOnField(error.getRegion().clazz, error.getRegion().member))
-            .filter(Objects::nonNull)
-            .collect(Collectors.toSet());
-
-    Set<AddAnnotation> suppressWarningsAnnotations =
-        fieldsWithSuppressWarnings.stream()
-            .map(
-                onField ->
-                    new AddSingleElementAnnotation(onField, "SuppressWarnings", "NullAway", false))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(suppressWarningsAnnotations);
-    // Update log.
-    config.log.updateInjectedAnnotations(suppressWarningsAnnotations);
-
-    // Collect NullAway.Init SuppressWarnings
-    Set<AddAnnotation> initializationSuppressWarningsAnnotations =
-        remainingFixes.stream()
-            .filter(
-                fix ->
-                    fix.isOnField()
-                        && (fix.reasons.contains("METHOD_NO_INIT")
-                            || fix.reasons.contains("FIELD_NO_INIT")))
-            // Filter nodes annotated with SuppressWarnings("NullAway")
-            .filter(fix -> !fieldsWithSuppressWarnings.contains(fix.toField()))
-            .map(
-                fix ->
-                    new AddSingleElementAnnotation(
-                        fix.toField(), "SuppressWarnings", "NullAway.Init", false))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(initializationSuppressWarningsAnnotations);
-    // Update log.
-    config.log.updateInjectedAnnotations(initializationSuppressWarningsAnnotations);
-    // Collect @NullUnmarked annotations on classes for any remaining error.
-    Utility.buildTarget(config);
-    remainingErrors = Utility.readErrorsFromOutputDirectory(config, config.targetModuleContext);
-    nullUnMarkedAnnotations =
-        remainingErrors.stream()
-            .filter(error -> !error.getRegion().isInAnonymousClass())
-            .map(
-                error ->
-                    new AddMarkerAnnotation(
-                        config.targetModuleContext.getLocationOnClass(error.getRegion().clazz),
-                        config.nullUnMarkedAnnotation))
-            .collect(Collectors.toSet());
-    injector.injectAnnotations(nullUnMarkedAnnotations);
-    // Update log.
-    config.log.updateInjectedAnnotations(nullUnMarkedAnnotations);
+  private void suppressRemainingErrors() {
+    Set<AddAnnotation> suppressionAnnotations = new NullAway().getSuppressionAnnotations(config);
+    config.log.updateInjectedAnnotations(suppressionAnnotations);
+    this.injector.injectAnnotations(suppressionAnnotations);
   }
 }
