+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/Checker.java
@@ -24,8 +24,8 @@
 
 package edu.ucr.cs.riple.core;
 
-import edu.ucr.cs.riple.core.io.deserializers.CheckerDeserializer;
-import edu.ucr.cs.riple.core.io.deserializers.nullaway.NullAwayV3Deserializer;
+import edu.ucr.cs.riple.core.checkers.CheckerDeserializer;
+import edu.ucr.cs.riple.core.checkers.deserializers.nullaway.NullAwayV3Deserializer;
 import javax.annotation.Nullable;
 
 /** Enum class to represent the supported checkers. */
