+++ annotator-core/src/main/java/edu/ucr/cs/riple/core/ModuleInfo.java
@@ -31,10 +31,10 @@ import java.nio.file.Paths;
 import java.util.Objects;
 import org.json.simple.JSONObject;
 
-/** Container class to hold paths to nullaway and scanner config files. */
+/** Container class to hold paths to checker and scanner config files. */
 public class ModuleInfo {
-  /** Path to nullaway config. */
-  public final Path nullawayConfig;
+  /** Path to checker config. */
+  public final Path checkerConfig;
   /** Path to scanner config. */
   public final Path scannerConfig;
   /** Directory where all serialized data from checkers are located. */
@@ -44,23 +44,23 @@ public class ModuleInfo {
    * Creates an instance of {@link ModuleInfo} from the given json object.
    *
    * @param id Global unique id for this module.
-   * @param globalDir Global path for all Annotator/Scanner/NullAway outputs.
+   * @param globalDir Global path for all Annotator/Scanner/Checker outputs.
    * @param jsonObject Json Object to retrieve values.
    * @return An instance of {@link ModuleInfo}.
    */
   public static ModuleInfo buildFromJson(int id, Path globalDir, JSONObject jsonObject) {
-    String nullawayConfigPath = (String) jsonObject.get("NULLAWAY");
+    String checkerConfigPath = (String) jsonObject.get("CHECKER");
     String scannerConfigPath = (String) jsonObject.get("SCANNER");
-    if (nullawayConfigPath == null || scannerConfigPath == null) {
+    if (checkerConfigPath == null || scannerConfigPath == null) {
       throw new IllegalArgumentException(
-          "Both paths to NullAway and Scanner config files must be set with NULLAWAY and SCANNER keys!");
+          "Both paths to Checker and Scanner config files must be set with CHECKER and SCANNER keys!");
     }
     return new ModuleInfo(
-        id, globalDir, Paths.get(nullawayConfigPath), Paths.get(scannerConfigPath));
+        id, globalDir, Paths.get(checkerConfigPath), Paths.get(scannerConfigPath));
   }
 
-  public ModuleInfo(int id, Path globalDir, Path nullawayConfig, Path scannerConfig) {
-    this.nullawayConfig = nullawayConfig;
+  public ModuleInfo(int id, Path globalDir, Path checkerConfig, Path scannerConfig) {
+    this.checkerConfig = checkerConfig;
     this.scannerConfig = scannerConfig;
     this.dir = globalDir.resolve(String.valueOf(id));
     try {
@@ -80,11 +80,11 @@ public class ModuleInfo {
       return false;
     }
     ModuleInfo that = (ModuleInfo) o;
-    return nullawayConfig.equals(that.nullawayConfig) && scannerConfig.equals(that.scannerConfig);
+    return checkerConfig.equals(that.checkerConfig) && scannerConfig.equals(that.scannerConfig);
   }
 
   @Override
   public int hashCode() {
-    return Objects.hash(nullawayConfig, scannerConfig);
+    return Objects.hash(checkerConfig, scannerConfig);
   }
 }
